feat(generic bp model): Move IBaseBusinessPartnerDto and dependent interfaces/classes to common
