+++ bpdm-common/src/main/kotlin/org/eclipse/tractusx/bpdm/common/dto/AddressType.kt
@@ -17,9 +17,12 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package org.eclipse.tractusx.bpdm.gate.api.model
+package org.eclipse.tractusx.bpdm.common.dto
 
-enum class BusinessPartnerRole {
-    SUPPLIER,
-    CUSTOMER
+enum class AddressType {
+
+    LegalAndSiteMainAddress,
+    LegalAddress,
+    SiteMainAddress,
+    AdditionalAddress
 }
