+++ bpdm-common/src/main/kotlin/org/eclipse/tractusx/bpdm/common/dto/BusinessPartnerRole.kt
@@ -17,22 +17,9 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package org.eclipse.tractusx.bpdm.gate.api.model
+package org.eclipse.tractusx.bpdm.common.dto
 
-import io.swagger.v3.oas.annotations.media.Schema
-
-@Schema(
-    description = "Identifier record for a business partner",
-    requiredProperties = ["type", "value"]
-)
-data class BusinessPartnerIdentifierDto(
-
-    @get:Schema(description = "Technical key of the type to which this identifier belongs to")
-    val type: String,
-
-    @get:Schema(description = "Value of the identifier")
-    val value: String,
-
-    @get:Schema(description = "Body which issued the identifier")
-    val issuingBody: String?
-)
+enum class BusinessPartnerRole {
+    SUPPLIER,
+    CUSTOMER
+}
