--- bpdm-gate-api/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/api/model/IBaseBusinessPartnerPostalAddressDto.kt
@@ -1,40 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021,2023 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ******************************************************************************/
-
-package org.eclipse.tractusx.bpdm.gate.api.model
-
-import io.swagger.v3.oas.annotations.media.Schema
-import org.eclipse.tractusx.bpdm.common.dto.IBaseAlternativePostalAddressDto
-import org.eclipse.tractusx.bpdm.common.dto.IBasePhysicalPostalAddressDto
-import org.eclipse.tractusx.bpdm.common.dto.openapidescription.LogisticAddressDescription
-import org.eclipse.tractusx.bpdm.gate.api.model.openapidescription.BusinessPartnerAddressDescription
-
-interface IBaseBusinessPartnerPostalAddressDto {
-
-    @get:Schema(description = BusinessPartnerAddressDescription.addressType)
-    val addressType: AddressType?
-
-    // TODO OpenAPI description for complex field does not work!!
-    @get:Schema(description = LogisticAddressDescription.physicalPostalAddress)
-    val physicalPostalAddress: IBasePhysicalPostalAddressDto?
-
-    // TODO OpenAPI description for complex field does not work!!
-    @get:Schema(description = LogisticAddressDescription.alternativePostalAddress)
-    val alternativePostalAddress: IBaseAlternativePostalAddressDto?
-}
