+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/entity/Roles.kt
@@ -39,8 +39,8 @@
 package org.eclipse.tractusx.bpdm.gate.entity
 
 import jakarta.persistence.*
+import org.eclipse.tractusx.bpdm.common.dto.BusinessPartnerRole
 import org.eclipse.tractusx.bpdm.common.model.BaseEntity
-import org.eclipse.tractusx.bpdm.gate.api.model.BusinessPartnerRole
 
 @Entity
 @Table(name = "roles")
